+++ src/Twilio/Rest/Video/V1/Room/RecordingRulesInstance.php
@@ -16,8 +16,6 @@ use Twilio\Values;
 use Twilio\Version;
 
 /**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- *
  * @property string $roomSid
  * @property string[] $rules
  * @property \DateTime $dateCreated