+++ src/Twilio/Rest/Video/V1/RoomOptions.php
@@ -29,10 +29,11 @@ abstract class RoomOptions {
      * @param string[] $videoCodecs An array of the video codecs that are supported
      *                              when publishing a track in the room
      * @param string $mediaRegion The region for the media server in Group Rooms
+     * @param array $recordingRules A collection of Recording Rules
      * @return CreateRoomOptions Options builder
      */
-    public static function create(bool $enableTurn = Values::NONE, string $type = Values::NONE, string $uniqueName = Values::NONE, string $statusCallback = Values::NONE, string $statusCallbackMethod = Values::NONE, int $maxParticipants = Values::NONE, bool $recordParticipantsOnConnect = Values::NONE, array $videoCodecs = Values::ARRAY_NONE, string $mediaRegion = Values::NONE): CreateRoomOptions {
-        return new CreateRoomOptions($enableTurn, $type, $uniqueName, $statusCallback, $statusCallbackMethod, $maxParticipants, $recordParticipantsOnConnect, $videoCodecs, $mediaRegion);
+    public static function create(bool $enableTurn = Values::NONE, string $type = Values::NONE, string $uniqueName = Values::NONE, string $statusCallback = Values::NONE, string $statusCallbackMethod = Values::NONE, int $maxParticipants = Values::NONE, bool $recordParticipantsOnConnect = Values::NONE, array $videoCodecs = Values::ARRAY_NONE, string $mediaRegion = Values::NONE, array $recordingRules = Values::ARRAY_NONE): CreateRoomOptions {
+        return new CreateRoomOptions($enableTurn, $type, $uniqueName, $statusCallback, $statusCallbackMethod, $maxParticipants, $recordParticipantsOnConnect, $videoCodecs, $mediaRegion, $recordingRules);
     }
 
     /**
@@ -66,8 +67,9 @@ class CreateRoomOptions extends Options {
      * @param string[] $videoCodecs An array of the video codecs that are supported
      *                              when publishing a track in the room
      * @param string $mediaRegion The region for the media server in Group Rooms
+     * @param array $recordingRules A collection of Recording Rules
      */
-    public function __construct(bool $enableTurn = Values::NONE, string $type = Values::NONE, string $uniqueName = Values::NONE, string $statusCallback = Values::NONE, string $statusCallbackMethod = Values::NONE, int $maxParticipants = Values::NONE, bool $recordParticipantsOnConnect = Values::NONE, array $videoCodecs = Values::ARRAY_NONE, string $mediaRegion = Values::NONE) {
+    public function __construct(bool $enableTurn = Values::NONE, string $type = Values::NONE, string $uniqueName = Values::NONE, string $statusCallback = Values::NONE, string $statusCallbackMethod = Values::NONE, int $maxParticipants = Values::NONE, bool $recordParticipantsOnConnect = Values::NONE, array $videoCodecs = Values::ARRAY_NONE, string $mediaRegion = Values::NONE, array $recordingRules = Values::ARRAY_NONE) {
         $this->options['enableTurn'] = $enableTurn;
         $this->options['type'] = $type;
         $this->options['uniqueName'] = $uniqueName;
@@ -77,6 +79,7 @@ class CreateRoomOptions extends Options {
         $this->options['recordParticipantsOnConnect'] = $recordParticipantsOnConnect;
         $this->options['videoCodecs'] = $videoCodecs;
         $this->options['mediaRegion'] = $mediaRegion;
+        $this->options['recordingRules'] = $recordingRules;
     }
 
     /**
@@ -181,6 +184,17 @@ class CreateRoomOptions extends Options {
      */
     public function setMediaRegion(string $mediaRegion): self {
         $this->options['mediaRegion'] = $mediaRegion;
+        return $this;
+    }
+
+    /**
+     * A collection of Recording Rules that describe how to include or exclude matching tracks for recording
+     *
+     * @param array $recordingRules A collection of Recording Rules
+     * @return $this Fluent Builder
+     */
+    public function setRecordingRules(array $recordingRules): self {
+        $this->options['recordingRules'] = $recordingRules;
         return $this;
     }