+++ tests/Twilio/Integration/Bulkexports/V1/Export/ExportCustomJobTest.php
@@ -80,7 +80,9 @@ class ExportCustomJobTest extends HolodeckTestCase {
                         "end_day": "end_day",
                         "webhook_url": "webhook_url",
                         "email": "email",
-                        "resource_type": "resource_type"
+                        "resource_type": "resource_type",
+                        "job_queue_position": "1",
+                        "estimated_completion_time": "2021-03-15T20:20:14.547"
                     }
                 ]
             }
@@ -125,7 +127,9 @@ class ExportCustomJobTest extends HolodeckTestCase {
                 "end_day": "end_day",
                 "webhook_url": "webhook_url",
                 "email": "email",
-                "resource_type": "resource_type"
+                "resource_type": "resource_type",
+                "job_queue_position": "1",
+                "estimated_completion_time": "2021-03-15T20:20:14.547"
             }
             '
         ));